+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `listener::set_up_vector` now Takes `Vector3f` instead of `&Vector3f`
 - Use `usize` instead of `u32` for `CustomShapePoints`
 - Use `usize` instead of `u32` for `Shape` methods
+- Use `usize` instead of `u32` in `CircleShape` and `ConvexShape` methods
 
 ### Internal improvements
 - Remove needless raw conversion methods for Rect